Coordinate CSPM and delta-SPH coefficient computation
We can compute the delta-SPH renormalized gradients as part of the CSPM
computations also used for the CCSPH formulation, without needing to
save FCOEFF to the device global memory and then reload it in a
different kernel. Indeed, if using delta-SPH without CCSPH, we don't
need to store FCOEFF at all.

Achieve this by

* moving the previous renormalized density gradient computation into a
  function template that is called by the CSPM kernel if appropriate;
* similarly isolate the thresholding for the FCOEFF storage in CSPM for
  CCSPH to a separate function template, to be called if appropriate;
* avoid early exit from the tensor computation loop if using delta-SPH.