Internally rename DELTA_SPH to ANTUONO, keep the alias
As explained in the comment in particledefine.h, this is because
delta-SPH proper refers to a whole scheme, whose density diffusion term
is the one devised by Antuono et al. as a correction to Molteni &
Colagrossi. However, at least as far as density diffusion is concerned,
the names refer to the same thing, so we keep DELTA_SPH as an alias,
allowing users to choose the diffusion term of DELTA_SPH even without
knowing that it's effectively the Antuono one.