Fix VTK array size for split arrays
When saving directly arrays with 4 components as an (optional) array
with 3 and an (optional) array of 1 component without conversions,
the on-disk size was erroneously computed from the 4-vector data type
rather than from the single-component data type, resulting in invalid
.vtp files. This went unnoticed until now due to the presence of
additional arrays afterwards that tapered over the bug. Spotted and
reported by JoJo in
https://gpusph.discourse.group/t/failed-when-open-the-vtp-file/160/9
and
https://github.com/GPUSPH/gpusph/issues/74

Fixes #74.