OpenMP usage fixes, take 2
No, you cannot use `#pragma omp reduction()` without an actual OpenMP
directive. The correct way to do this is using the `scope` directive.

(The only reason why it seemed to work was that g++ ignored the pragma
when it didn't understand it, without any warning or error. Clang
errors out, and the warning can be seen with g++ by enabling
-Wnknown-pragmas.)

Problem is, the `scope` directive is an OpenMP 5.1 pragma —and GCC
versions up to 11 do not support it anyway, nor Clang up to version 12.

So in practice OpenMP support for the CPU backend is only possible
with GCC 12 or who knows which future version of Clang.
(You can build with older GCC versions, but the reductions may give the
wrong results.)