Streamline Chrono usage in headers
Since we're declaring the needed Chrono types without importing the
Chrono headers, we can use them throughout GPUSPH, without having to
change the function signatures depending on whether Chrono is defined or
not. This allows us to eliminate a bunch of USE_CHRONO preprocessor
fencing.

(More could probably be eliminated, this is just a quick pass as needed
for the next commit.)