Separate FEA exchange buffers
In preparation for the split of the FEA_STEP command, separate the
BUFFER_FEA_EXCH buffer into a FORCES and a VEL buffer.

This will prevents the VEL buffer from being clobbered during forces
computation, and should allow us to avoid the extra transfers
between host and device to keep the correct values on it.