Script to plot a single testpoint
Apparently, extracting individual testpoint data with ParaView is a bit
of a pain, due to the approach we use to save one instant per file
instead of one testpoint per file. Until we get more customizable
writers, this simple script can be used to extract the data for a single
testpoint and either plot it (if the script stdout is the terminal, it's
automatically redirected to gnuplot) or print it out (so that it can be
redirected to file).