New neighbors list iterator structure
Instead of simple iterators for each type, all deriving from a core
iterator with the required variables, and a multi-iterator inheriting
from the simple ones, we design the multi-type iterator in a recursive
fashion, with the next iterator as base class of the previous one,
and the PT_NONE iterator (optionally assimilated to the core iterator)
acting as the terminator for the list.

This leads to a massive simplification of the code base, and most
importantly it eliminates the virtual inheritance for the simple
iterators, which was apparently a big obstacle to optimizations both
with nvcc and some Clang versions (but not others ... e.g. Clang 12
managed to produce better code than nvcc or Clang 13).

The net result is a massive gain in performance regardless of compilers,
with the largest gain being with nvcc, which jumps from the bottom to
the top.