Further simplification of the iterator code
Give the core an iterator-like interface (reset() and next()) to
avoid special-casing the nested iterator code for the “last iterator”
case.

Also clean up the user-facing type by using an auxiliary structure to
remove PT_NONE from the type list and to map the empty type list to the
core.