Installation guide in AsciiDoctor format
This is not a direct conversion of the LaTeX sources, but written from
scratch, with only the essential information.

The additional information in the old LaTeX installation guide will be
moved to appropriate sections of the user guide, with the exception of
sections we do not intent to maintain (e.g. how to build Chrono and the
now-unsupported support for GPUSPH in external tools such as Crixus and
SALOME).