Get started on the new user guide
Some of the content is new or completely rewritten,
other parts are ported over from the old LaTeX user guide.

At this commit, we imported all of the contents from gpusph-install.tex,
with the exception of the instructions about SALOME and Crixus
(which we do not support anymore),
and about how to install Chrono and CUDA (which we deem not to be our
responsibility anymore).