Refactor neibsInCell
We can leverage the fact that particles within cell are sorted by type
to refactor the neighbors search function into a type-dependent template
that is called for each type in sequence.

The compile-time knowledge of the type we're processing allows, among
other things, to eliminate the dynamic indexing of the neibs_num array,
and thus its transfer from local memory to registers.