buildneibs: cell-based kernel execution
On GPU, the buildneibs kernels is plagued by uncoalesced globabl memory
accesses, due to massive scattering of data fetches.

To try and reduce this issue, on GPU we map work-groups to cells instead
of work-items to particles, as this should leave to more consistent
exploration of neighboring cells.

Note that there is an intrisic inefficiency in this approach, because
with the usual range of values for the kernel radius and smoothing
lenght, cells typically have less than 32 particles each, so in most
cases part of the work-group will be masked even with a block size of
32.