Cache neighbor data on shared memory when building the neighbors list
When using cell-based buildNeibsListDevice, we can leverage the fact
that all work-items in a work-group go through the same neighboring
cells to improve memory access patterns.

The idea is that, when traversing a neighboring cell, we load the
info and pos of the particles in that cell in shared memory
(each work-item loads one). We then fetch the data from the
shared-memory cache instead of global memory.

This improves coalescence in accessing the global memory, and leverages
the broadcasting capabilities of shared memory.

Quick testing shows improvements around 30% over the original
particle-based neighbors list construction.