buildneibs: warp-based cell mapping
Instead of a cell per block, run with a _warp_ per block, but grouping
multiple warps in a block. This has two purposes: reduce the number of
blocks, and potentially increase the number of cells being processed.

The latter is due to some GPUs having a max blocks per SM limit which
is lower than the max warps per SM. E.g., my GTX 1650 has a blocks per
SM limit of 16, and a warps per SM limit of 32, so it's necessary to
have at least two warps per block to achieve maximum occupancy.