Torus: fix bounding box
The intent was to provide a cubic bounding box that overestimated the
actual bounding box, but was valid regardless of the rotation of the
torus. However, it failed to take into account that the maximum particle
distance from the center is the sum of both radii, and it failed to
double the distance when computing the wrapping cube, with the result
that the result only covered an octant of the torus if even that.

The fix is to use twice the sum of the radii as cube side, and to take
as corner the origin shifted by half that. The resulting bounding box is
overestimated, but at least it does what it's supposed to do.