API 1: improve handling of domain size for multi-layer boundaries
We used to add a margin equal to the inter-particle spacing times ONE
LESS than the number of layers. This is too low in the case of tangent
filling, even with the extra half deltap that we added
to help in periodic cases.

Instead, add m_deltap/2 along periodic directions and the full
layers*deltap in the non-periodic ones. This may be half a deltap too
large in the BORDER_CENTERED filling case, but we can live with that.