World dimensions should be set as early as possible
Setting the world dimensions in the ProblemCore initialize() member
function may be too late, particularly if the Problem (API 1) calls
(indirectly) any of the dimension-dependent geometry functions, such as
Volume() for (n-1)-dimensional geometries. This is the case for example
for setParticleMassByDensity().

The solution is to set the Object world dimensions as soon as the
dimensionality is known, i.e. at SimParams creation time. Since we don't
want to include Object.h everywhere simparams.h is included, the call is
moved to a separate implementation file.